+++ packages/dotmac-files/src/dotmac_files/manifest.py
@@ -10,7 +10,7 @@ from dotmac_files.models import PLATFORM_TABLES, TENANT_TABLES
 
 module = ModuleManifest(
     code="files",
-    version="0.1.0a1",
+    version="0.1.0a2",
     core=False,
     short_code="files",
     migration_prefix="fi",